CR fixes: leaflet-crs util, follow DDAU in base-layer Changed leaflet-wfst dependency to tag-specific
